+++ src/components/user-profile-head.jsx
@@ -45,10 +45,16 @@ import { UserProfileHeadActions } from './user-profile-head-actions';
 import { UserProfileHeadStats } from './user-profile-head-stats';
 import { lazyComponent } from './lazy-component';
 
-const UserSubscriptionEditPopup = lazyComponent(() => import('./user-subscription-edit-popup'), {
-  fallback: 'Loading form...',
-  errorMessage: "Couldn't load form",
-});
+const UserSubscriptionEditPopup = lazyComponent(
+  () =>
+    import('./user-subscription-edit-popup').then((m) => ({
+      default: m.UserSubscriptionEditPopup,
+    })),
+  {
+    fallback: 'Loading form...',
+    errorMessage: "Couldn't load form",
+  },
+);
 
 export const UserProfileHead = withRouter(
   withKey(({ router }) => router.params.userName)(function UserProfileHead({ router }) {
@@ -397,9 +403,11 @@ function InListsIndicator({
           ))}
         </>
       )}{' '}
+      (
       <ButtonLink ref={subscrFormPivotRef} onClick={subscrFormToggle}>
         edit
       </ButtonLink>
+      )
     </div>
   );
 }